+++ src/Contracts/MigratableInterface.php
@@ -24,25 +24,4 @@ interface MigratableInterface
      * @return string
      */
     public function getMigrationPrefix(): string;
-
-    /**
-     * Register module migrations.
-     *
-     * @return void
-     */
-    public function registerMigrations(): void;
-
-    /**
-     * Run module migrations.
-     *
-     * @return void
-     */
-    public function runMigrations(): void;
-
-    /**
-     * Rollback module migrations.
-     *
-     * @return void
-     */
-    public function rollbackMigrations(): void;
 }